+++ notes/changelog.md
@@ -20,6 +20,7 @@ Add all relevant changes to either ADDED, CHANGED, REMOVED or FIXED along with t
 - [10/6/2024] @kambains226  fixed the game move error , created a splash screen with an input
 - [13/6/2024] @kambains226  fixed the error when the user clicks the switch and display the score to the screen 
 - [14/6/2024] @kambains226  added the guesses to the stats screen and implement a restart button 
+- [14/6/2024] @stasyanski   fixed retrieval from localstorage when null
 
 ### Changed
 
@@ -27,6 +28,7 @@ Add all relevant changes to either ADDED, CHANGED, REMOVED or FIXED along with t
 - [12/6/2024] @stasyanski   changed and better positioned the heart in lives mode
 - [12/6/2024] @stasyanski   reworking the game over splash
 - [12/6/2024] @stasyanski   updated the lives and score css to be more consistent
+- [14/6/2024] @stasyanski   reused the settings css for stats screen
 
 ### Removed